Split views into separate files corresponding the pages they handle
