+++ WebCLI/views/__init__.py
@@ -1,4 +1,6 @@
 # flake8: noqa
 
-from .views import *
+from .my_algorithms import *
 from .algorithm_details_view import algorithm_details_view
+from .homepage import *
+from .new_version import *
