+++ WebCLI/views/new_algorithm.py
@@ -1,0 +1,24 @@
+from django.shortcuts import render
+from django.contrib.auth.decorators import login_required
+from django.utils import timezone
+from ..models import Algorithm, Algorithm_version
+from ..forms import AlgorithmForm
+from ..forms import AlgorithmVersionForm
+
+
+@login_required
+def new_algorithm(request):
+    aform = AlgorithmForm(initial={'user': request.user})
+    vform = AlgorithmVersionForm()
+    if request.method == "POST":
+        algorithm_form = AlgorithmForm(request.POST)
+        new_algorithm = algorithm_form.save(commit=False)
+        new_algorithm.user = request.user
+        new_algorithm.save()
+        algorithm = AlgorithmVersionForm(request.POST).data['algorithm']
+        v = Algorithm_version(timestamp=timezone.now(), algorithm_id=new_algorithm,
+                              algorithm=algorithm)
+        v.save()
+    data = {'algorithms': Algorithm.objects.filter(user=request.user), 'aform': aform,
+            'vform': vform}
+    return render(request, 'WebCLI/newAlgorithm.html', data)
