+++ WebCLI/views/new_algorithm_type.py
@@ -1,0 +1,14 @@
+from django.shortcuts import render
+from django.contrib.auth.decorators import login_required
+from ..models import Algorithm_type
+from ..forms import AlgorithmTypeForm
+
+
+@login_required
+def new_algorithm_type(request):
+    form = AlgorithmTypeForm()
+    if request.method == "POST":
+        m = AlgorithmTypeForm(request.POST)
+        m.save()
+    data = {'types': Algorithm_type.objects.all(), 'form': form}
+    return render(request, 'WebCLI/newAlgorithmType.html', data)
